More progress on multithread...
Unfortunately multithreading seems like a lost cause. Since threads
share the same address space, they share the same breakpoints. This
means when a thread starts executing a function, we can't remove the
start breakpoint for that function because another thread might want to
execute the function at the same time. Even if we single-step after the
breakpoint and immediately put it back there is a race condition and
another thread might still make it through without being caught. The
only way would be to stop all threads whenever we find a breakpoint,
remove the breakpoint, single-step for that one thread, put the
breakpoint back, and then resume all threads. This seems like a hassle
and a performance problem. Not sure if it's really worth it.